Modify car.py and used_cars.py in prac_06 feedback
