+++ prac_06/car.py
@@ -5,3 +5,33 @@ This program uses Car class as an example
 Estimate: 100 minutes
 Actual:    minutes
 """
+
+
+class Car:
+    """Represent a Car object."""
+
+    def __int__(self, fuel=0):
+        """Initialise a Car instance.
+
+        fuel: float, one unit of fuel drives one kilometre
+        """
+        self.fuel = fuel
+        self._odometer = 0
+
+    def add_fuel(self, amount):
+        """Add amount to the car's fuel"""
+        self.fuel += amount
+
+    def drive(self, distance):
+        """Drive the car a given distance.
+
+        Drive given distance if car has enough fuel
+        or drive until fuel runs out return the distance actually driven.
+        """
+        if distance > self.fuel:
+            distance = self.fuel
+            self.fuel = 0
+        else:
+            self.fuel -= distance
+        self._odometer += distance
+        return distance
